% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_args}
\alias{check_args}
\title{Check arguments to function}
\usage{
check_args(..., expand = FALSE)
}
\arguments{
\item{...}{Numerics. Truncated Weibull parameters.}

\item{expand}{Logical. Whether to expand arguments to same length (default: FALSE).}
}
\value{
list
}
\description{
This function does some basic checks of the arguments going
into one of the truncated Weibull functions. It then returns
a list containing those arguments.
}
\keyword{internal}
