% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtweibull.R
\name{dtweibull}
\alias{dtweibull}
\title{Mass function of truncated Weibull distribution}
\usage{
dtweibull(x, shape, scale = 1, a = 0, b = Inf, log = FALSE)
}
\arguments{
\item{x}{Numeric, value at which to evaluate mass function}

\item{shape}{Numeric, Weibull shape parameter}

\item{scale}{Numeric, Weibull scale parameter}

\item{a}{Numeric, lower truncation point}

\item{b}{Numeric, upper truncation point}

\item{log}{Logical, whether to return log mass function value}
}
\value{
Numeric value
}
\description{
Mass function of truncated Weibull distribution
}
\examples{
dtweibull(3, 0.5, 1.5, 2)
}
