% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htweibull.R
\name{Htweibull}
\alias{Htweibull}
\title{Truncated Weibull cumulative hazard}
\usage{
Htweibull(x, shape, scale = 1, a = 0, b = Inf, log = FALSE)
}
\arguments{
\item{x}{(Numerics) Place to evaluate cumulative hazard}

\item{shape}{(Numerics) Weibull shape parameter}

\item{scale}{(Numerics) Weibull scale parameter (default: 1)}

\item{a}{(Numerics) Lower truncation point (default: 0)}

\item{b}{(Numerics) Upper truncation point (default: Inf)}

\item{log}{(Logical) Return log-cumhaz (default: FALSE)}
}
\value{
(Numerics) Value of cumulative hazard
}
\description{
Truncated Weibull cumulative hazard
}
\examples{
Htweibull(3, 2.5, 1.5, 1, 5)
}
