% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ldenom}
\alias{ldenom}
\title{Compute Denominator for a Truncated Weibull Distribution}
\usage{
ldenom(a, b, shape, scale)
}
\arguments{
\item{a}{Lower interval endpoint}

\item{b}{Upper interval endpoint}

\item{shape}{Shape parameter}

\item{scale}{Scale parameter}
}
\value{
A numeric vector
}
\description{
This is an internal function.
}
\keyword{internal}
