% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_select}
\alias{list_select}
\title{Select Elements from List of Vectors}
\usage{
list_select(..., x, ind)
}
\arguments{
\item{...}{Names arguments to make into a list}

\item{x}{List to select elements from}

\item{ind}{Logical indicators of which elements to choose}
}
\value{
Modified list
}
\description{
This function selects specified elements from vectors contained in a list.
For vectors of length 1 it just returns that value. The assumption when
this was written is that all vectors have a common length or length 1.
}
\keyword{internal}
