% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{log1mpexp}
\alias{log1mpexp}
\title{Numerically compute the value of log(1 - p + p * exp(-x))}
\usage{
log1mpexp(log_p, a, b, shape, scale, ...)
}
\arguments{
\item{a}{(Numeric) Lower truncation point}

\item{b}{(Numeric) Upper truncation point}

\item{shape}{(Numeric) Weibull shape parameter}

\item{scale}{(Numeric) Weibull scale parameter}

\item{...}{Arguments that get passed along to `uniroot`}

\item{p}{(Numeric) Numeric 0 <= p <= 1}
}
\value{
A numeric value
}
\description{
Due to numerical accuracy issues with directly computing this
function, a different approach is used that is thought
to be more robust. The approach uses uniroot, and therefore
may fail. This is an internal function to the pacakge and no
error checking is performed on its inputs.
}
\keyword{interval}
