% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_args}
\alias{make_args}
\title{Check arguments to function}
\usage{
make_args(...)
}
\arguments{
\item{...}{Numerics. Truncated Weibull parameters.}
}
\value{
list
}
\description{
This function does some basic checks of the arguments going
into one of the truncated Weibull functions. It then returns
a list containing those arguments.
}
