% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptweibull.R
\name{ptweibull}
\alias{ptweibull}
\title{Cumulative Distribution Function for Truncated Weibull}
\usage{
ptweibull(
  q,
  shape,
  scale = 1,
  a = 0,
  b = Inf,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{Numeric, value at which to evaluate CDF}

\item{shape}{Numeric, Weibull shape parameter}

\item{scale}{Numeric, Weibull scale parameter}

\item{a}{Numeric, lower truncation point}

\item{b}{Numeric, upper truncation point}

\item{lower.tail}{Logical, whether to compute lower or upper tail probabilities}

\item{log.p}{Logical, whether to return log probabilities}
}
\value{
Numeric value(s)
}
\description{
Cumulative Distribution Function for Truncated Weibull
}
\examples{
ptweibull(2, shape=1.5, a=1)
}
