% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtweibull.R
\name{qtweibull}
\alias{qtweibull}
\title{Quantile Function for Truncated Weibull}
\usage{
qtweibull(p, a = 0, b = Inf, shape, scale, log.p = FALSE, ...)
}
\arguments{
\item{p}{(Numeric) Vector of probabilities}

\item{a}{(Numeric) Vector of lower truncation points}

\item{b}{(Numeric) Vector of upper truncation points}

\item{shape}{(Numeric) Vector of Weibull shape parameters}

\item{scale}{(Numeric) Vector of Weibull scale parameters}

\item{log.p}{(Logical) Whether p represents p or \eqn{log(p)}}

\item{...}{}
}
\value{
Numeric vector of quantile function values
}
\description{
This is an attempt to provide a more numerically stable
version of the truncated Weibull distribution, especially
for more extreme truncation points. Underneith the hood it
relies on \code{uniroot}, which may issue errors. This should
be considered the least stable of the functions associated
with this distribution in this package.
}
\examples{
qtweibull(0.5, 1, 5, 1.5, 2.5)
}
