% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htweibull.R
\name{htweibull}
\alias{htweibull}
\title{Truncated Weibull hazard function}
\usage{
htweibull(x, shape, scale = 1, a = 0, b = Inf, log = FALSE)
}
\arguments{
\item{x}{Numerics. Value at which to evaluate hazard function.}

\item{shape}{Numerics. Shape parameter.}

\item{scale}{Numerics. Scale parameter (default: 1).}

\item{a}{Numerics. Lower truncation point (default: 0).}

\item{b}{Numerics. Upper truncation point (default: Inf).}

\item{log}{Logical. Return log-hazard (default: FALSE).}
}
\value{
Numerics. Hazard value.
}
\description{
Truncated Weibull hazard function
}
\examples{
htweibull(5, 2.5)
}
