% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htweibull.R
\name{htweibull}
\alias{htweibull}
\alias{Htweibull}
\title{Truncated Weibull hazard function}
\usage{
htweibull(x, shape, scale = 1, a = 0, b = Inf, log = FALSE)

Htweibull(x, shape, scale = 1, a = 0, b = Inf, log = FALSE)
}
\arguments{
\item{x}{Numerics. Value at which to evaluate hazard or cumulative hazard function.}

\item{shape}{Numerics. Shape parameter.}

\item{scale}{Numerics. Scale parameter (default: 1).}

\item{a}{Numerics. Lower truncation point (default: 0).}

\item{b}{Numerics. Upper truncation point (default: Inf).}

\item{log}{Logical. Return log-hazard or log-cumulative hazard (default: FALSE).}
}
\value{
Numerics. Value of hazard or cumulative hazard function.
}
\description{
Truncated Weibull hazard function

Truncated Weibull cumulative hazard
}
\examples{
htweibull(5, 2.5)
Htweibull(3, 2.5, 1.5, 1, 5)
}
